pyproject.toml
[project]
name = "napari-afmslicer"
dynamic = ["version"]
description = "A simple plugin to use AFMSlicer with napari"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
  {name = "Neil Shephard"},
  {email = "n.shephard@sheffield.ac.uk"},
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: napari",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.10"
# napari can be included in dependencies if napari imports are required.
# However, you should not include napari[all], napari[qt],
# or any other Qt bindings directly (e.g. PyQt5, PySide2).
# See best practices: https://napari.org/stable/plugins/building_a_plugin/best_practices.html
dependencies = [
    "afmslicer @ git+https://github.com/AFM-SPM/AFMSlicer",
    "napari-afmreader",
    "napari-crop",
    "napari-skimage",
    "napari-topostats",
    "numpy",
    "magicgui",
    "qtpy",
    "scikit-image",
]

[project.optional-dependencies]
# Allow easily installation with the full, default napari installation
# (including Qt backend) using napari-afmslicer[all].
all = ["napari[all]"]
tests = [
  "pytest",
  "pytest-cov",
  "pytest-mpl",
  "pytest-xdist",
  "syrupy",
]
dev = [
  "napari-afmslicer[tests]",
  "black",
  "codespell",
  "ipdb",
  "mypy",
  "numpydoc",
  "pre-commit",
  "pydocstyle[toml]",
  "pylint",
  "pytest-testmon",
  "ruff",
  "tomli",
]

[dependency-groups]
dev = [
    "black",
    "codespell",
    "ipdb",
    "mypy",
    "napari[qt]",  # test with napari's default Qt bindings
    "numpydoc",
    "pre-commit",
    "pydocstyle[toml]",
    "pylint",
    "pytest",  # https://docs.pytest.org/en/latest/contents.html
    "pytest-cov",  # https://pytest-cov.readthedocs.io/en/latest/
    "pytest-mpl",
    "pytest-qt",  # https://pytest-qt.readthedocs.io/en/latest/
    "pytest-xdist",
    "pytest-testmon",
    "ruff",
    "syrupy",
    "tomli",
    "tox-uv",
]

[project.entry-points."napari.manifest"]
napari-afmslicer = "napari_afmslicer:napari.yaml"

[project.urls]
"Bug Tracker" = "https://github.com/AFM-SPM/napari-afmslicer/issues"
"Documentation" = "https://github.com/AFM-SPM/napari-afmslicer#README.md"
"Source Code" = "https://github.com/AFM-SPM/napari-afmslicer"
"User Support" = "https://github.com/AFM-SPM/napari-afmslicer/issues"

[build-system]
requires = ["setuptools>=42.0.0", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[tool.setuptools_scm]
write_to = "src/napari_afmslicer/_version.py"
fallback_version = "0.0.1+nogit"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
# filterwarnings = ["error"]  # recommended by PP309; deprecation warnings cause failures so leaving for the user to enable if desired
log_cli_level = "INFO"
testpaths = ["tests"]

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.ruff]
line-length = 120
lint.select = [
    "E", "F", "W", #flake8
    "UP", # pyupgrade
    "I", # isort
    "BLE", # flake8-blind-exception
    "B", # flake8-bugbear
    "A", # flake8-builtins
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "G", # flake8-logging-format
    "PIE", # flake8-pie
    "SIM", # flake8-simplify
]
lint.ignore = [
    "E501", # line too long. let black handle this
    "UP006", "UP007", # type annotation. As using magicgui require runtime type annotation then we disable this.
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "*vendored*",
    "*_vendor*",
]
fix = true

# Numpy2 deprecation checks
lint.extend-select = ["NPY201"]
preview = false

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.isort]
case-sensitive = true

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = true

[tool.codespell]
skip = '*.spm*,*.mplstyle,*.svg,*.stp,*.top,*.002,*.004'
count = ''
quiet-level = 3
ignore-words-list = 'OT'

[tool.numpydoc_validation]
checks = [
  "all", # Perform all check except those listed below
  "ES01",
  "EX01",
  "PR10", # Conflicts with black formatting
  "SA01",
]
exclude = [  # don't report on objects that match any of these regex
    "\\.undocumented_method$",
    "\\.__repr__$",
    "^test_",
    "^conftest",
    "^conf$",
    "^notebooks/*.py$"
]